requirements.txt
aioredis==1.3.1
aliyun-python-sdk-core==2.13.25
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-rds==2.5.1
amqp==5.0.1
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
cffi==1.14.2
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.3
conf==0.4.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
daphne==2.5.0
Django==3.1.2
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-filter==2.3.0
django-js-asset==1.2.2
django-redis==4.12.1
django-rest-swagger==2.2.0
django-taggit==1.3.0
django-tinymce==3.0.2
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
inflection==0.5.1
isort==5.5.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
mysql-connector-python==8.0.21
openapi-codec==1.3.2
packaging==20.4
Pillow==7.2.0
prompt-toolkit==3.0.8
protobuf==3.13.0
py3Fdfs==2.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==0.10.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytz==2020.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
zope.interface==5.1.0
